# Move some hash computations to compile time