--- a/src/StableHashTraits.jl
+++ b/src/StableHashTraits.jl
@@ -298,6 +298,7 @@ function stable_hash_helper(obj, hash_state::MarkerHash{<:BufferedHash}, context
 end
 
 function stable_hash_helper(obj, hash_state::BufferedHash, context, ::WriteHash)
+    @show obj
     write(hash_state.io, obj, context)
     flush_bytes!(hash_state)
     return hash_state
@@ -701,13 +702,13 @@ function hash_method(::Type, c::HashVersion{1})
     return (ConstantHash("Base.DataType"), FnHash(qualified_type_))
 end
 function hash_method(::Type, c::HashVersion)
-    return (ConstantHash(@inthash("Base.DataType")), FnHash(stable_type_id))
+    return (ConstantHash(@inthash("Base.DataType")), TypeHash(c))
 end
 function hash_method(::Function, c::HashVersion{1})
     return (ConstantHash("Base.Function"), FnHash(qualified_name_))
 end
 function hash_method(::Function, c::HashVersion)
-    return (ConstantHash(@inthash("Base.Function")), FnHash(stable_typename_id))
+    return (ConstantHash(@inthash("Base.Function")), TypeHash(c))
 end
 function hash_method(::AbstractSet, c::HashVersion)
     return (TypeNameHash(c), FnHash(sort! ∘ collect))