--- a/test/references/ref10_2_sha1.txt
+++ b/test/references/ref10_2_sha1.txt
@@ -1,1 +1,1 @@
-56b53bce2d5be7bedd58a56ba220bea5c72ddabc
+6fc38d114768eec3657bffa9c184feb1342ba36a