… .github/pull_request_template.md
… README.md
… coverage.lcov
… test/benchmark_records.md
… test/references/ref00_2_crc.txt
… test/references/ref00_2_crc32c.txt
… test/references/ref00_2_sha1.txt
… test/references/ref00_2_sha256.txt
… test/references/ref01_2_crc.txt
… test/references/ref01_2_crc32c.txt
… test/references/ref01_2_sha1.txt
… test/references/ref01_2_sha256.txt
… test/references/ref02_2_crc.txt
… test/references/ref02_2_crc32c.txt
… test/references/ref02_2_sha1.txt
… test/references/ref02_2_sha256.txt
… test/references/ref03_2_crc.txt
… test/references/ref03_2_crc32c.txt
… test/references/ref03_2_sha1.txt
… test/references/ref03_2_sha256.txt
… test/references/ref04_2_crc.txt
… test/references/ref04_2_crc32c.txt
… test/references/ref04_2_sha1.txt
… test/references/ref04_2_sha256.txt
… test/references/ref05_2_crc.txt
… test/references/ref05_2_crc32c.txt
… test/references/ref05_2_sha1.txt
… test/references/ref05_2_sha256.txt
… test/references/ref06_2_crc.txt
… test/references/ref06_2_crc32c.txt