@haberdashPI 2023-10-18
> One other question I have is why the reference test files needed to change if the default HashVersion{1} is still being used and these changes are non-breaking?

Just to address this before I take a look through your other comments. 

I have reference tests for each HashVersion: I want to test that given a particular hash version the hash value does not change for a given object. The function `test_hash` is defined on lines 31-32 of `runtests.jl` such that it will use the context set in the for loop.