@haberdashPI 2023-10-18
To elaborate on my comment above: in the reference tests, `ref01_2_crc` is reference test 1, using `HashVersion{2}` with a `crc32c` hash function. You will find that none of the reference tests matching `ref[nun]_1_[hash]` have changed, confirming that no `HashVersion{1}` return values have changed in the tests.